Done for now, moving on with the course
Left the exercises on 2.18, started to implement the show button on the countries lists but there are some values that aren't reachable to parts that can manipulate the code better that the country displayer component. Also, it sends multiple requests, probably by not using useEffect. I've done so much from now and I'd prefer to advance to lessons in part 3